+++ bindings/builtinserver/builtinserver_posix.go
@@ -1,9 +1,10 @@
 // +build !windows
-//go:generate sh -c "cd ../.. && mkdir -p build && cd build && cmake -DGO_BUILTIN_SERVER_EXPORT_PKG_PATH=\"../../bindings/builtinserver\" .. && make reindexer_server_library reindexer -j8"
+//go:generate sh -c "cd ../.. && mkdir -p build && cd build && cmake -DLINK_RESOURCES=On -DGO_BUILTIN_SERVER_EXPORT_PKG_PATH=\"../../bindings/builtinserver\" .. && make reindexer_server_library reindexer -j4"
 
 package builtinserver
 
-// #cgo CXXFLAGS: -std=c++11 -g -O2 -Wall -Wpedantic -Wextra -I../../cpp_src 
-// #cgo CFLAGS: -std=c99 -g -O2 -Wall -Wpedantic -Wno-unused-variable -I../../cpp_src 
-// #cgo LDFLAGS: -L${SRCDIR}/../../build/cpp_src/ -L${SRCDIR}/../../build/cpp_src/server/ -lreindexer_server_library  -lreindexer -lleveldb -lsnappy -ldl -lreindexer_server_resources -lm -lstdc++ -g
+// #cgo pkg-config: libreindexer_server
+// #cgo CXXFLAGS: -std=c++11 -g -O2 -Wall -Wpedantic -Wextra 
+// #cgo CFLAGS: -std=c99 -g -O2 -Wall -Wpedantic -Wno-unused-variable 
+// #cgo LDFLAGS:  -g
 import "C"